// bushub-client/frontend/src/components/common/DeviceListShowDetail/types.ts
export type DeviceCardVariant = 'default' | 'panel';
// ...
export interface DeviceListShowDetailHandle {
  handleFormChange: (key: string, value: any, deviceId?: string, unitId?: string) => void;
}